validate did for sao update permission
