+++ x/sao/types/expected_keepers.go
@@ -76,6 +76,7 @@ type ModelKeeper interface {
 type DidKeeper interface {
 	GetCosmosPaymentAddress(ctx sdk.Context, did string) (sdk.AccAddress, error)
 	GetSidDocument(ctx sdk.Context, versionId string) (val types2.SidDocument, found bool)
+	ValidDid(ctx sdk.Context, did string) error
 }
 
 // MarketKeeper
